TEST: Pass valid external URL to test payload
To create a realistic test scenario the implementation can work
with, we have to provide a valid external testing URL.

Creating URLs for localhost is bound to fail because of
the missing webserver configuration on testing environments.
Furthermore, we are not aiming to test resource delivery from the local
web host in this case.